+++ scripts/create_neodos_image.py
@@ -200,7 +200,7 @@ Happy hacking!
         # ── Read binary data ──
         userbin_dir = os.path.join(os.path.dirname(__file__), '..', 'userbin')
         nxe_files = {}
-        for name in ['cpuinfo', 'neoshell', 'neoinit', 'coredir', 'cd', 'corehelp', 'datetime', 'ver', 'mem', 'vol', 'echo', 'kobj', 'coretype', 'tree']:
+        for name in ['cpuinfo', 'neoshell', 'neoinit', 'coredir', 'cd', 'corehelp', 'datetime', 'ver', 'mem', 'vol', 'echo', 'kobj', 'coretype', 'tree', 'corecls', 'corecopy', 'coredel', 'coreren', 'coremd', 'corerd']:
             fpath = os.path.join(userbin_dir, f'{name}.nxe')
             data = b''
             if os.path.exists(fpath):
@@ -286,9 +286,10 @@ Happy hacking!
         # 14=Libraries, 15-18,44=NXL files
         # 19=Layouts, 20-21=NKB files
         # 22=Config, 23-24=cfg files
-        # 25=Programs, 26-36=NXE files
+        # 25=Programs, 26-36,45-53=NXE files
         # 37=Packages, 38=Users, 39=Default, 40=Alejandro
         # 41=Temp, 42=Data, 43=Logs, 46=type.nxe, 47=tree.nxe
+        # 48=cls.nxe, 49=copy.nxe, 50=del.nxe, 51=ren.nxe, 52=md.nxe, 53=rd.nxe
 
         # Read NEM driver data
         nem_data = {
@@ -316,6 +317,12 @@ Happy hacking!
         kobj_blocks       = alloc_blocks(45, len(nxe_files['kobj']))
         coretype_blocks   = alloc_blocks(46, len(nxe_files['coretype']))
         tree_blocks       = alloc_blocks(47, len(nxe_files['tree']))
+        corecls_blocks    = alloc_blocks(48, len(nxe_files['corecls']))
+        corecopy_blocks   = alloc_blocks(49, len(nxe_files['corecopy']))
+        coredel_blocks    = alloc_blocks(50, len(nxe_files['coredel']))
+        coreren_blocks    = alloc_blocks(51, len(nxe_files['coreren']))
+        coremd_blocks     = alloc_blocks(52, len(nxe_files['coremd']))
+        corerd_blocks     = alloc_blocks(53, len(nxe_files['corerd']))
         fs_nxl_blocks     = alloc_blocks(15, len(nxl_data))
         cpuinfo_nxl_blocks2= alloc_blocks(18, len(cpuinfo_nxl_data))
         math_nxl_blocks   = alloc_blocks(44, len(math_nxl_data))
@@ -336,7 +343,7 @@ Happy hacking!
         lib_dir_blocks    = alloc_blocks(14, 1536)   # Libraries dir (5 entries)
         lay_dir_blocks    = alloc_blocks(19, 768)    # Layouts dir (2 entries + padding)
         cfg_dir_blocks    = alloc_blocks(22, 768)    # Config dir (2 entries + padding)
-        prog_dir_blocks   = alloc_blocks(25, 4096)   # Programs dir (16 entries)
+        prog_dir_blocks   = alloc_blocks(25, 5120)   # Programs dir (20 entries)
         pkg_dir_blocks    = alloc_blocks(37, 256)    # Packages dir (empty)
         usr_dir_blocks    = alloc_blocks(38, 768)    # Users dir
         def_dir_blocks    = alloc_blocks(39, 256)    # Users\Default (empty)
@@ -361,7 +368,7 @@ Happy hacking!
             22: (dir_mode, 768,  pad_blocks(cfg_dir_blocks)),
             23: (MODE_FILE | default_perms_for_filename("system.cfg"), len(system_cfg_content), pad_blocks(system_cfg_blocks)),
             24: (MODE_FILE | default_perms_for_filename("input.cfg"), len(input_cfg_content), pad_blocks(input_cfg_blocks)),
-            25: (dir_mode, 3840, pad_blocks(prog_dir_blocks)),
+            25: (dir_mode, 5120, pad_blocks(prog_dir_blocks)),
             26: (MODE_FILE | default_perms_for_filename("NeoShell.nxe"), len(nxe_files['neoshell']), pad_blocks(neoshell_blocks)),
             27: (MODE_FILE | default_perms_for_filename("NeoInit.nxe"), len(nxe_files['neoinit']), pad_blocks(neoinit_blocks)),
             28: (MODE_FILE | default_perms_for_filename("cpuinfo.nxe"), len(nxe_files['cpuinfo']), pad_blocks(cpuinfo_blocks)),
@@ -374,8 +381,14 @@ Happy hacking!
             34: (MODE_FILE | default_perms_for_filename("vol.nxe"), len(nxe_files['vol']), pad_blocks(vol_blocks)),
             35: (MODE_FILE | default_perms_for_filename("echo.nxe"), len(nxe_files['echo']), pad_blocks(echo_blocks)),
              45: (MODE_FILE | default_perms_for_filename("kobj.nxe"), len(nxe_files['kobj']), pad_blocks(kobj_blocks)),
-             46: (MODE_FILE | default_perms_for_filename("type.nxe"), len(nxe_files['coretype']), pad_blocks(coretype_blocks)),
-             47: (MODE_FILE | default_perms_for_filename("tree.nxe"), len(nxe_files['tree']), pad_blocks(tree_blocks)),
+              46: (MODE_FILE | default_perms_for_filename("type.nxe"), len(nxe_files['coretype']), pad_blocks(coretype_blocks)),
+              47: (MODE_FILE | default_perms_for_filename("tree.nxe"), len(nxe_files['tree']), pad_blocks(tree_blocks)),
+              48: (MODE_FILE | default_perms_for_filename("cls.nxe"), len(nxe_files['corecls']), pad_blocks(corecls_blocks)),
+              49: (MODE_FILE | default_perms_for_filename("copy.nxe"), len(nxe_files['corecopy']), pad_blocks(corecopy_blocks)),
+              50: (MODE_FILE | default_perms_for_filename("del.nxe"), len(nxe_files['coredel']), pad_blocks(coredel_blocks)),
+              51: (MODE_FILE | default_perms_for_filename("ren.nxe"), len(nxe_files['coreren']), pad_blocks(coreren_blocks)),
+              52: (MODE_FILE | default_perms_for_filename("md.nxe"), len(nxe_files['coremd']), pad_blocks(coremd_blocks)),
+              53: (MODE_FILE | default_perms_for_filename("rd.nxe"), len(nxe_files['corerd']), pad_blocks(corerd_blocks)),
              37: (dir_mode, 256,  pad_blocks(pkg_dir_blocks)),
             38: (dir_mode, 768,  pad_blocks(usr_dir_blocks)),
             39: (dir_mode, 256,  pad_blocks(def_dir_blocks)),
@@ -555,6 +568,12 @@ Happy hacking!
         image[offset+2816:offset+3072]= create_dir_entry(45, 1, "kobj.nxe")
         image[offset+3072:offset+3328]= create_dir_entry(46, 1, "type.nxe")
         image[offset+3328:offset+3584]= create_dir_entry(47, 1, "tree.nxe")
+        image[offset+3584:offset+3840]= create_dir_entry(48, 1, "cls.nxe")
+        image[offset+3840:offset+4096]= create_dir_entry(49, 1, "copy.nxe")
+        image[offset+4096:offset+4352]= create_dir_entry(50, 1, "del.nxe")
+        image[offset+4352:offset+4608]= create_dir_entry(51, 1, "ren.nxe")
+        image[offset+4608:offset+4864]= create_dir_entry(52, 1, "md.nxe")
+        image[offset+4864:offset+5120]= create_dir_entry(53, 1, "rd.nxe")
 
         # Write all NXE binary data
         nxe_inode_map = {
@@ -572,6 +591,12 @@ Happy hacking!
             45: ('kobj.nxe', nxe_files['kobj']),
              46: ('type.nxe', nxe_files['coretype']),
              47: ('tree.nxe', nxe_files['tree']),
+             48: ('cls.nxe', nxe_files['corecls']),
+             49: ('copy.nxe', nxe_files['corecopy']),
+             50: ('del.nxe', nxe_files['coredel']),
+             51: ('ren.nxe', nxe_files['coreren']),
+             52: ('md.nxe', nxe_files['coremd']),
+             53: ('rd.nxe', nxe_files['corerd']),
         }
         for inum, (name, data) in nxe_inode_map.items():
             if not data: